+++ src/test/groovy/ru/vyarus/dropwizard/guice/config/debug/renderer/StatRenderWithDwBundleTest.groovy
@@ -10,7 +10,7 @@ import ru.vyarus.dropwizard.guice.diagnostic.BaseDiagnosticTest
 import ru.vyarus.dropwizard.guice.diagnostic.support.bundle.FooBundle
 import ru.vyarus.dropwizard.guice.diagnostic.support.features.FooModule
 import ru.vyarus.dropwizard.guice.diagnostic.support.features.FooResource
-import ru.vyarus.dropwizard.guice.module.context.debug.DiagnosticBundle
+import ru.vyarus.dropwizard.guice.module.GuiceyConfigurationInfo
 import ru.vyarus.dropwizard.guice.module.context.debug.report.stat.StatsRenderer
 import ru.vyarus.dropwizard.guice.module.installer.feature.LifeCycleInstaller
 import ru.vyarus.dropwizard.guice.test.spock.UseGuiceyApp
@@ -25,7 +25,12 @@ import javax.inject.Inject
 class StatRenderWithDwBundleTest extends BaseDiagnosticTest {
 
     @Inject
+    GuiceyConfigurationInfo info
     StatsRenderer renderer
+
+    void setup() {
+        renderer = new StatsRenderer(info)
+    }
 
     def "Check guicey app stats render"() {
         /* render would look like:
@@ -49,7 +54,7 @@ class StatRenderWithDwBundleTest extends BaseDiagnosticTest {
     │   └── 3 extensions recognized from 7 classes in 6.967 ms
     │
     ├── [42%] INJECTOR created in 263.6 ms
-    │   ├── from 6 guice modules
+    │   ├── from 5 guice modules
     │   │
     │   └── injector log
     │       ├── Module execution: 178 ms
@@ -86,7 +91,7 @@ class StatRenderWithDwBundleTest extends BaseDiagnosticTest {
 
         render.contains("] BUNDLES")
         render.contains("2 resolved in")
-        render.contains("7 initialized in")
+        render.contains("6 initialized in")
         render.contains("1 dropwizard bundles initialized in")
 
         render.contains("] COMMANDS")
@@ -97,7 +102,7 @@ class StatRenderWithDwBundleTest extends BaseDiagnosticTest {
         render.contains("3 extensions recognized from 7 classes in")
 
         render.contains("] INJECTOR")
-        render.contains("from 6 guice modules")
+        render.contains("from 5 guice modules")
         render.contains("injector log")
         render.contains("Module execution")
 
@@ -126,7 +131,7 @@ class StatRenderWithDwBundleTest extends BaseDiagnosticTest {
                             .searchCommands()
                             .dropwizardBundles(new DBundle())
                             .bundles(new FooBundle())
-                            .modules(new FooModule(), new DiagnosticBundle.DiagnosticModule())
+                            .modules(new FooModule())
                             .disableInstallers(LifeCycleInstaller)
                             .strictScopeControl()
                             .printDiagnosticInfo()