+++ src/test/groovy/ru/vyarus/dropwizard/guice/config/unique/DuplicateBundlesReportingTest.groovy
@@ -6,6 +6,7 @@ import io.dropwizard.setup.Bootstrap
 import io.dropwizard.setup.Environment
 import ru.vyarus.dropwizard.guice.AbstractTest
 import ru.vyarus.dropwizard.guice.GuiceBundle
+import ru.vyarus.dropwizard.guice.module.GuiceyConfigurationInfo
 import ru.vyarus.dropwizard.guice.module.context.debug.report.diagnostic.DiagnosticConfig
 import ru.vyarus.dropwizard.guice.module.context.debug.report.diagnostic.DiagnosticRenderer
 import ru.vyarus.dropwizard.guice.module.context.debug.report.tree.ContextTreeConfig
@@ -24,10 +25,14 @@ import javax.inject.Inject
 class DuplicateBundlesReportingTest extends AbstractTest {
 
     @Inject
+    GuiceyConfigurationInfo info
     DiagnosticRenderer renderer
-    @Inject
     ContextTreeRenderer treeRenderer
 
+    void setup() {
+        renderer = new DiagnosticRenderer(info)
+        treeRenderer = new ContextTreeRenderer(info)
+    }
 
     def "Check diagnostic info render"() {
 
@@ -37,7 +42,6 @@ class DuplicateBundlesReportingTest extends AbstractTest {
     BUNDLES =
         Bundle                       (r.v.d.g.c.u.DuplicateBundlesReportingTest) *REG(5/10)
         MiddleBundle                 (r.v.d.g.c.u.DuplicateBundlesReportingTest)
-        DiagnosticBundle             (r.v.d.g.m.c.debug)
         HK2DebugBundle               (r.v.d.g.m.j.debug)        *HOOK
         GuiceRestrictedConfigBundle  (r.v.d.g.support.util)     *HOOK
         CoreInstallersBundle         (r.v.d.g.m.installer)
@@ -52,7 +56,6 @@ class DuplicateBundlesReportingTest extends AbstractTest {
     GUICE MODULES =
         HK2DebugModule               (r.v.d.g.m.j.d.HK2DebugBundle)
         GRestrictModule              (r.v.d.g.s.u.GuiceRestrictedConfigBundle)
-        DiagnosticModule             (r.v.d.g.m.c.d.DiagnosticBundle)
         GuiceBootstrapModule         (r.v.d.guice.module)
 """ as String;
     }
@@ -63,7 +66,6 @@ class DuplicateBundlesReportingTest extends AbstractTest {
         cleanupReport(treeRenderer.renderReport(new ContextTreeConfig())) == """
 
     APPLICATION
-    ├── module     DiagnosticModule             (r.v.d.g.m.c.d.DiagnosticBundle)
     ├── module     GuiceBootstrapModule         (r.v.d.guice.module)
     ├── Bundle                       (r.v.d.g.c.u.DuplicateBundlesReportingTest)
     ├── -Bundle                      (r.v.d.g.c.u.DuplicateBundlesReportingTest) *IGNORED(2)
@@ -75,8 +77,6 @@ class DuplicateBundlesReportingTest extends AbstractTest {
     │   ├── Bundle#4                     (r.v.d.g.c.u.DuplicateBundlesReportingTest)
     │   ├── -Bundle#4                    (r.v.d.g.c.u.DuplicateBundlesReportingTest) *IGNORED(2)
     │   └── Bundle#5                     (r.v.d.g.c.u.DuplicateBundlesReportingTest)
-    │
-    ├── DiagnosticBundle             (r.v.d.g.m.c.debug)
     │
     ├── CoreInstallersBundle         (r.v.d.g.m.installer)
     │   ├── installer  -JerseyFeatureInstaller      (r.v.d.g.m.i.f.jersey)     *IGNORED