+++ src/test/groovy/ru/vyarus/dropwizard/guice/config/unique/DuplicateDropwizardBundlesReportingTest.groovy
@@ -7,6 +7,7 @@ import io.dropwizard.setup.Bootstrap
 import io.dropwizard.setup.Environment
 import ru.vyarus.dropwizard.guice.AbstractTest
 import ru.vyarus.dropwizard.guice.GuiceBundle
+import ru.vyarus.dropwizard.guice.module.GuiceyConfigurationInfo
 import ru.vyarus.dropwizard.guice.module.context.debug.report.diagnostic.DiagnosticConfig
 import ru.vyarus.dropwizard.guice.module.context.debug.report.diagnostic.DiagnosticRenderer
 import ru.vyarus.dropwizard.guice.module.context.debug.report.tree.ContextTreeConfig
@@ -25,10 +26,14 @@ import javax.inject.Inject
 class DuplicateDropwizardBundlesReportingTest extends AbstractTest {
 
     @Inject
+    GuiceyConfigurationInfo info
     DiagnosticRenderer renderer
-    @Inject
     ContextTreeRenderer treeRenderer
 
+    void setup() {
+        renderer = new DiagnosticRenderer(info)
+        treeRenderer = new ContextTreeRenderer(info)
+    }
 
     def "Check diagnostic info render"() {
 
@@ -38,7 +43,6 @@ class DuplicateDropwizardBundlesReportingTest extends AbstractTest {
     BUNDLES =
         DBundle                      (r.v.d.g.c.u.DuplicateDropwizardBundlesReportingTest) *DW, REG(7/12)
         MiddleBundle                 (r.v.d.g.c.u.DuplicateDropwizardBundlesReportingTest)
-        DiagnosticBundle             (r.v.d.g.m.c.debug)
         HK2DebugBundle               (r.v.d.g.m.j.debug)        *HOOK
         GuiceRestrictedConfigBundle  (r.v.d.g.support.util)     *HOOK
         CoreInstallersBundle         (r.v.d.g.m.installer)
@@ -53,7 +57,6 @@ class DuplicateDropwizardBundlesReportingTest extends AbstractTest {
     GUICE MODULES =
         HK2DebugModule               (r.v.d.g.m.j.d.HK2DebugBundle)
         GRestrictModule              (r.v.d.g.s.u.GuiceRestrictedConfigBundle)
-        DiagnosticModule             (r.v.d.g.m.c.d.DiagnosticBundle)
         GuiceBootstrapModule         (r.v.d.guice.module)
 """ as String;
     }
@@ -64,7 +67,6 @@ class DuplicateDropwizardBundlesReportingTest extends AbstractTest {
         cleanupReport(treeRenderer.renderReport(new ContextTreeConfig())) == """
 
     APPLICATION
-    ├── module     DiagnosticModule             (r.v.d.g.m.c.d.DiagnosticBundle)
     ├── module     GuiceBootstrapModule         (r.v.d.guice.module)
     ├── DBundle                      (r.v.d.g.c.u.DuplicateDropwizardBundlesReportingTest) *DW
     ├── -DBundle                     (r.v.d.g.c.u.DuplicateDropwizardBundlesReportingTest) *DW, IGNORED(2)
@@ -81,8 +83,6 @@ class DuplicateDropwizardBundlesReportingTest extends AbstractTest {
     │   │
     │   ├── -DBundle#5                   (r.v.d.g.c.u.DuplicateDropwizardBundlesReportingTest) *DW, IGNORED(2)
     │   └── DBundle#7                    (r.v.d.g.c.u.DuplicateDropwizardBundlesReportingTest) *DW
-    │
-    ├── DiagnosticBundle             (r.v.d.g.m.c.debug)
     │
     ├── CoreInstallersBundle         (r.v.d.g.m.installer)
     │   ├── installer  -JerseyFeatureInstaller      (r.v.d.g.m.i.f.jersey)     *IGNORED